fix: correct create board feature
